Move implementation of I2cFirmata back to cpp
The rationale given for keeping it in the .h file seems outdated
